fix: address github-advanced-security (jscpd) review comments
一覧のタイトル取得が codex 側と重複していた（jscpd が 60 トークンの重複を検出）。
`transcript-head.ts` に切り出し、codex 側も同じものを使うようにして重複を実際に消した。

- `parseJsonRecord` — 完全に同一だった
- `cleanTitle` — 既定タイトルだけが違ったので引数にした
- open / read / stat / close の並び — `readTranscriptHead` に集約。mtime を同じ
  ハンドルから取る理由（追記中のファイルは read と stat の間に消え得る）をコメントに残した
